FIX/Fix: Move SAT constraints for AND in correct module
